Replace size_t with int32_t so that buf_t is definitely larger than word_t and addr_t

+++ src/emu_types.h
@@ -17,11 +17,10 @@
 #define TYPES_H
 
 #include <stdint.h>
-#include <stdlib.h>
 
 typedef uint8_t word_t;
 typedef uint16_t addr_t;
-typedef size_t buf_t; 
+typedef uint32_t buf_t; 
 
 #define HALF_WORD_SIZE (4)
 #define HALF_ADDR_SIZE (8)
